fix import statements and file organization - update page object imports to match renamed files, remove duplicate combo product file, fix test imports across all spec files
