+++ tests/turf kingAcelepryn.spec.ts
@@ -1,40 +1,46 @@
 import { test, expect } from '@playwright/test';
-import { TurfKingBifenthrinPage } from '../pages/turf king Bifenthrin';
+import { TurfKingAceleprynPage } from '../pages/turfkingAcelepryn';
 
-test.describe('Turf King Bifenthrin Page Tests', () => {
-    let turfkingbifenthrinpage: TurfKingBifenthrinPage;
+test.describe('Turf King Acelepryn Page Tests', () => {
+    let turfkingaceleprynpage: TurfKingAceleprynPage;
 
     test.beforeEach(async ({ page }) => {
-        turfkingbifenthrinpage = new TurfKingBifenthrinPage(page);
+        turfkingaceleprynpage = new TurfKingAceleprynPage(page);
     });
 
-    test('verify the turf king bifenthrin page elements and functions', async ({ page }) => {
+    test('verify the contents of the turf king acelepryn page', async ({ page }) => {
         // Set longer timeout for this test
-        test.setTimeout(120000); // 2 minutes
+        test.setTimeout(300000); // 5 minutes
         
         // Navigate to home page
         await page.goto('/');
         await page.waitForLoadState('networkidle');
         await page.waitForTimeout(1000);
         
-        // Navigate through the flow to reach Bifenthrin page
-        await turfkingbifenthrinpage.clickTurfProductsTitle();
+        // Navigate through the flow to reach Acelepryn page
+        await turfkingaceleprynpage.clickTurfProductsTitle();
         await page.waitForLoadState('networkidle');
         await page.waitForTimeout(1000);
         
-        await turfkingbifenthrinpage.clickCombinationProductLearnMoreButton();
+        await turfkingaceleprynpage.clickCombinationProductLearnMoreButton();
         await page.waitForLoadState('networkidle');
         await page.waitForTimeout(1000);
         
-        await turfkingbifenthrinpage.clickTurfKingBifenthrinLearnMoreButton();
+        await turfkingaceleprynpage.clickTurfKingAceleprynLearnMoreButton();
         await page.waitForLoadState('networkidle');
         await page.waitForTimeout(1000);
         
         // Verify we're on the correct page
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
         
         // Wait for page elements to load
         await page.waitForTimeout(1000);
+        
+        // Verify page elements are visible
+        await expect(turfkingaceleprynpage.turfkingaceleprynheading).toBeVisible();
+        await expect(turfkingaceleprynpage.turfkingaceleprynpagedescription).toBeVisible();   
+        await expect(turfkingaceleprynpage.aceleprynimagealttext).toBeVisible();
+        await expect(turfkingaceleprynpage.spreadersettingbutton).toBeVisible();
         
         // Add wait before clicking spreader settings
         await page.waitForTimeout(500);
@@ -42,7 +48,7 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
         // Test spreader settings button
         const [pdfPage] = await Promise.all([
             page.context().waitForEvent('page'),
-            turfkingbifenthrinpage.clickSpreadersettingButton()
+            turfkingaceleprynpage.clickSpreadersettingButton()
         ]);
         
         // Wait for PDF with shorter timeout and check URL
@@ -54,20 +60,20 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
             // Even if URL doesn't match exactly, continue the test
         }
         await pdfPage.close();
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
         await page.waitForTimeout(500);
 
-        // Test Turf King 16-0-5 with Bifenthrin buttons
+        // Test Turf King 12-0-3 with Acelepryn buttons
         
         // Wait before testing product buttons
         await page.waitForTimeout(500);
         
         // Label button
-        await expect(turfkingbifenthrinpage.turfking1605withbifenthrinlabelbutton).toBeVisible();
+        await expect(turfkingaceleprynpage.turfking1203aceleprynlabelbutton).toBeVisible();
         await page.waitForTimeout(300);
         const [pdfPageLabel] = await Promise.all([
             page.context().waitForEvent('page'),
-            turfkingbifenthrinpage.clickTurfKing1605withBifenthrinLabelButton()
+            turfkingaceleprynpage.clickTurfKing1203AceleprynLabelButton()
         ]);
         try {
             await pdfPageLabel.waitForLoadState('load', { timeout: 5000 });
@@ -76,15 +82,15 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
             console.log('Label PDF URL:', await pdfPageLabel.url());
         }
         await pdfPageLabel.close();
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
         await page.waitForTimeout(500);
         
         // SDS button
-        await expect(turfkingbifenthrinpage.turfking1605withbifenthrinSDSbutton).toBeVisible();
+        await expect(turfkingaceleprynpage.turfking1203aceleprynSDSbutton).toBeVisible();
         await page.waitForTimeout(300);
         const [pdfPageSDS] = await Promise.all([
             page.context().waitForEvent('page'),
-            turfkingbifenthrinpage.clickTurfKing1605withBifenthrinSDSButton()
+            turfkingaceleprynpage.clickTurfKing1203AceleprynSDSButton()
         ]);
         try {
             await pdfPageSDS.waitForLoadState('load', { timeout: 5000 });
@@ -93,15 +99,15 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
             console.log('SDS PDF URL:', await pdfPageSDS.url());
         }
         await pdfPageSDS.close();
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
         await page.waitForTimeout(500);
         
         // Spec sheet button
-        await expect(turfkingbifenthrinpage.turfking1605withbifenthrinspecsheetbutton).toBeVisible();
+        await expect(turfkingaceleprynpage.turfking1203aceleprynspecsheetbutton).toBeVisible();
         await page.waitForTimeout(300);
         const [pdfPageSpec] = await Promise.all([
             page.context().waitForEvent('page'),
-            turfkingbifenthrinpage.clickTurfKing1605withBifenthrinSpecSheetButton()
+            turfkingaceleprynpage.clickTurfKing1203AceleprynSpecSheetButton()
         ]);
         try {
             await pdfPageSpec.waitForLoadState('load', { timeout: 5000 });
@@ -110,17 +116,36 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
             console.log('Spec PDF URL:', await pdfPageSpec.url());
         }
         await pdfPageSpec.close();
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
 
-        // Test Turf King 0-0-7 with Bifenthrin buttons
+        // Test Turf King 17-0-5 with Acelepryn buttons
+        await page.waitForTimeout(500);
+        
+        // Label button for 17-0-5
+        await expect(turfkingaceleprynpage.turfking1705aceleprynlabelbutton).toBeVisible();
+        await page.waitForTimeout(300);
+        const [pdfPageLabel1705] = await Promise.all([
+            page.context().waitForEvent('page'),
+            turfkingaceleprynpage.clickTurfKing1705AceleprynLabelButton()
+        ]);
+        try {
+            await pdfPageLabel1705.waitForLoadState('load', { timeout: 5000 });
+            await expect(pdfPageLabel1705).toHaveURL(/.*label/i);
+        } catch (error) {
+            console.log('1705 Label PDF URL:', await pdfPageLabel1705.url());
+        }
+        await pdfPageLabel1705.close();
+        await expect(page).toHaveURL(/.*acelepryn/);
+
+        // Test Turf King 0-0-7 with Acelepryn buttons
         await page.waitForTimeout(500);
         
         // Label button for 0-0-7
-        await expect(turfkingbifenthrinpage.turfking007withbifenthrinlabelbutton).toBeVisible();
+        await expect(turfkingaceleprynpage.turfking007aceleprynlabelbutton).toBeVisible();
         await page.waitForTimeout(300);
         const [pdfPageLabel007] = await Promise.all([
             page.context().waitForEvent('page'),
-            turfkingbifenthrinpage.clickTurfKing007withBifenthrinLabelButton()
+            turfkingaceleprynpage.clickTurfKing007AceleprynLabelButton()
         ]);
         try {
             await pdfPageLabel007.waitForLoadState('load', { timeout: 5000 });
@@ -129,6 +154,6 @@ test.describe('Turf King Bifenthrin Page Tests', () => {
             console.log('007 Label PDF URL:', await pdfPageLabel007.url());
         }
         await pdfPageLabel007.close();
-        await expect(page).toHaveURL(/.*bifenthrin/);
+        await expect(page).toHaveURL(/.*acelepryn/);
     });
 });
